env exemple

+++ index.js
@@ -63,13 +63,17 @@ require("dotenv").config();
 		const objeto = req.body;
 
 		if (!objeto || !objeto.nome || !objeto.imagemUrl) {
-			res.send("Objeto invalido");
+			res.send(
+				"Requisição inválida, certifique-se que tenha os campos nome e imagemUrl"
+			);
 			return;
 		}
 
-		const insertCount = await personagens.insertOne(objeto);
-		console.log(insertCount);
-		if (!insertCount) {
+		const result = await personagens.insertOne(objeto);
+
+		console.log(result);
+		//Se ocorrer algum erro com o mongoDb esse if vai detectar
+		if (result.acknowledged == false) {
 			res.send("Ocorreu um erro");
 			return;
 		}
@@ -77,29 +81,66 @@ require("dotenv").config();
 		res.send(objeto);
 	});
 
+	//[PUT] Atualizar personagem
 	app.put("/personagens/:id", async (req, res) => {
 		const id = req.params.id;
 		const objeto = req.body;
-		res.send(
-			await personagens.updateOne(
-				{
-					_id: ObjectId(id),
-				},
-				{
-					$set: objeto,
-				}
-			)
-		);
+
+		if (!objeto || !objeto.nome || !objeto.imagemUrl) {
+			res.send(
+				"Requisição inválida, certifique-se que tenha os campos nome e imagemUrl"
+			);
+			return;
+		}
+
+		const quantidadePersonagens = await personagens.countDocuments({
+			_id: ObjectId(id),
+		});
+
+		if (quantidadePersonagens !== 1) {
+			res.send("Personagem não encontrado");
+			return;
+		}
+
+		const result = await personagens.updateOne(
+			{
+				_id: ObjectId(id),
+			},
+			{
+				$set: objeto,
+			}
+			);
+			//console.log(result);
+			//Se acontecer algum erro no MongoDb, cai na seguinte valiadação
+			if(result.modifiedCount !== 1){
+				res.send("Ocorreu um erro ao atualizar o personagem")
+			}
+			res.send(await getPersonagemById(id));
 	});
 
+	//[DELETE] Deleta um personagem
 	app.delete("/personagens/:id", async (req, res) => {
 		const id = req.params.id;
+		//Retorna a quantidade de personagens com o filtro(Id) especificado
+		const quantidadePersonagens = await personagens.countDocuments({
+			_id: ObjectId(id),
+		});
+		//Checar se existe o personagem solicitado
+		if (quantidadePersonagens !== 1) {
+			res.send("Personagem não encontrao");
+			return;
+		}
+		//Deletar personagem
+		const result = await personagens.deleteOne({
+			_id: ObjectId(id),
+		});
+		//Se não con
+		if (result.deletedCount !== 1) {
+			res.send("Ocorreu um erro ao remover o personagem");
+			return;
+		}
 
-		res.send(
-			await personagens.deleteOne({
-				_id: ObjectId(id),
-			})
-		);
+		res.send("Personagem removido com sucesso!");
 	});
 
 	app.listen(port, () => {
